🔧📝💄🐛Minor changes about the config file and the arguments.
The default config file is now well formatted and will run on localhost.
The --help feedback has been improved.
To combine importance level now you might use , as a separator and not a |
This is because the pipes are recognized as the builtin.
Fixed a bug when you send an unknowed flag.